requirements.txt
fastapi
uvicorn[standard]
ffmpeg-python
python-multipart
requests
google-cloud-speech
google-cloud-storage
google-auth

# OCR
pytesseract
numpy
Pillow
opencv-python
